Added tests for incoming eth
